Authenticating Users: Login & Register Done.
Login & Register
- - - - X
In order to create a login and register functionality, we had to do a
few things.
Logic:
Since both the login and register requires us to use authorization
functionality, we wrapped that into a service, that returns the
interface object $firebaseAuth(), an interface for authorization.
AngularFire’s $firebaseAuth factory service object provides
authentication methods to angular apps.

UI:
	For the View of this authentication component, we created the
templates, login.html & register.html. It’s M and V are managed in a
controller(Auth, $state) auth.controller with a couple of properties
and methods it provides to the UI:
 User object - email and password(given to it by the UI via ng-model)
 Login & register -
$firebaseAuth’s:
$createUserWithEmailAndPassword method allows angular to use firebase’s
authentication service, which creates a uid for each user and links
that uid to profile info, email, password.
$signInWithEmailAndPassword method just gets you the uid’s information
to provide the user custom experience in the app.
Problem:
Don’t import anything two times(js becomes confused)
The order of dependency injections matters. (why does it matter though?
Why did switching the order of authCtrl fix the problem of auth’s
methods not being recognized?)

Next Step: decoupling.